backend/pyproject.toml
[project]
name = "backend-normal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.6",
    "fastapi[standard]>=0.115.12",
    "jwcrypto>=1.5.6",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "python-jose>=3.4.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
]
